Rename cosmos_client_components::components to ::impls
