+++ crates/cosmos/cosmos-client-components/src/impls/all.rs
@@ -39,40 +39,40 @@ use hermes_relayer_components::chain::traits::types::packet::IbcPacketTypesProvi
 use hermes_relayer_components::chain::traits::types::status::ChainStatusTypeComponent;
 use hermes_relayer_components::chain::traits::types::timestamp::TimestampTypeComponent;
 
-use crate::components::ack_packet_message::BuildCosmosAckPacketMessage;
-use crate::components::ack_packet_payload::BuildCosmosAckPacketPayload;
-use crate::components::channel_handshake_message::BuildCosmosChannelHandshakeMessage;
-use crate::components::channel_handshake_payload::BuildCosmosChannelHandshakePayload;
-use crate::components::connection_handshake_message::BuildCosmosConnectionHandshakeMessage;
-use crate::components::connection_handshake_payload::BuildCosmosConnectionHandshakePayload;
-use crate::components::create_client_message::BuildCosmosCreateClientMessage;
-use crate::components::create_client_payload::BuildCreateClientPayloadWithChainHandle;
-use crate::components::init_channel_options::ProvideCosmosInitChannelOptionsType;
-use crate::components::init_connection_options::ProvideCosmosInitConnectionOptionsType;
-use crate::components::packet_fields::CosmosPacketFieldReader;
-use crate::components::packet_from_ack::BuildCosmosPacketFromWriteAck;
-use crate::components::query_block::QueryCometBlock;
-use crate::components::query_chain_id::QueryChainIdWithChainHandle;
-use crate::components::query_chain_status::QueryChainStatusWithChainHandle;
-use crate::components::query_client_state::QueryCosmosClientStateFromChainHandle;
-use crate::components::query_consensus_state::QueryCosmosConsensusStateFromChainHandle;
-use crate::components::query_consensus_state_height::QueryConsensusStateHeightFromChainHandle;
-use crate::components::query_packet_commitments::QueryCosmosPacketCommitments;
-use crate::components::query_received_packet::QueryReceivedPacketWithChainHandle;
-use crate::components::query_send_packet::QueryCosmosSendPacket;
-use crate::components::query_send_packets::QuerySendPacketsConcurrently;
-use crate::components::query_unreceived_packet::QueryUnreceivedCosmosPacketSequences;
-use crate::components::query_write_ack_event::QueryWriteAckEventFromChainHandle;
-use crate::components::receive_packet_message::BuildCosmosReceivePacketMessage;
-use crate::components::receive_packet_payload::BuildCosmosReceivePacketPayload;
-use crate::components::send_messages_as_tx::SendMessagesToTxContext;
-use crate::components::timeout_packet_message::BuildCosmosTimeoutPacketMessage;
-use crate::components::timeout_packet_payload::BuildCosmosTimeoutPacketPayload;
-use crate::components::types::chain::ProvideCosmosChainTypes;
-use crate::components::types::connection_handshake_payload::ProvideCosmosConnectionHandshakePayloads;
-use crate::components::types::create_client_options::ProvideCosmosCreateClientSettings;
-use crate::components::update_client_message::BuildCosmosUpdateClientMessage;
-use crate::components::update_client_payload::BuildUpdateClientPayloadWithChainHandle;
+use crate::impls::ack_packet_message::BuildCosmosAckPacketMessage;
+use crate::impls::ack_packet_payload::BuildCosmosAckPacketPayload;
+use crate::impls::channel_handshake_message::BuildCosmosChannelHandshakeMessage;
+use crate::impls::channel_handshake_payload::BuildCosmosChannelHandshakePayload;
+use crate::impls::connection_handshake_message::BuildCosmosConnectionHandshakeMessage;
+use crate::impls::connection_handshake_payload::BuildCosmosConnectionHandshakePayload;
+use crate::impls::create_client_message::BuildCosmosCreateClientMessage;
+use crate::impls::create_client_payload::BuildCreateClientPayloadWithChainHandle;
+use crate::impls::init_channel_options::ProvideCosmosInitChannelOptionsType;
+use crate::impls::init_connection_options::ProvideCosmosInitConnectionOptionsType;
+use crate::impls::packet_fields::CosmosPacketFieldReader;
+use crate::impls::packet_from_ack::BuildCosmosPacketFromWriteAck;
+use crate::impls::query_block::QueryCometBlock;
+use crate::impls::query_chain_id::QueryChainIdWithChainHandle;
+use crate::impls::query_chain_status::QueryChainStatusWithChainHandle;
+use crate::impls::query_client_state::QueryCosmosClientStateFromChainHandle;
+use crate::impls::query_consensus_state::QueryCosmosConsensusStateFromChainHandle;
+use crate::impls::query_consensus_state_height::QueryConsensusStateHeightFromChainHandle;
+use crate::impls::query_packet_commitments::QueryCosmosPacketCommitments;
+use crate::impls::query_received_packet::QueryReceivedPacketWithChainHandle;
+use crate::impls::query_send_packet::QueryCosmosSendPacket;
+use crate::impls::query_send_packets::QuerySendPacketsConcurrently;
+use crate::impls::query_unreceived_packet::QueryUnreceivedCosmosPacketSequences;
+use crate::impls::query_write_ack_event::QueryWriteAckEventFromChainHandle;
+use crate::impls::receive_packet_message::BuildCosmosReceivePacketMessage;
+use crate::impls::receive_packet_payload::BuildCosmosReceivePacketPayload;
+use crate::impls::send_messages_as_tx::SendMessagesToTxContext;
+use crate::impls::timeout_packet_message::BuildCosmosTimeoutPacketMessage;
+use crate::impls::timeout_packet_payload::BuildCosmosTimeoutPacketPayload;
+use crate::impls::types::chain::ProvideCosmosChainTypes;
+use crate::impls::types::connection_handshake_payload::ProvideCosmosConnectionHandshakePayloads;
+use crate::impls::types::create_client_options::ProvideCosmosCreateClientSettings;
+use crate::impls::update_client_message::BuildCosmosUpdateClientMessage;
+use crate::impls::update_client_payload::BuildUpdateClientPayloadWithChainHandle;
 
 pub struct CosmosClientComponents;
