+++ crates/cosmos/cosmos-client-components/src/lib.rs
@@ -3,7 +3,7 @@
 
 extern crate alloc;
 
-pub mod components;
+pub mod impls;
 pub mod methods;
 pub mod traits;
 pub mod types;
