+++ crates/cosmos/cosmos-relayer/src/impls/chain/components/query_client_state.rs
@@ -1,5 +1,5 @@
 use cgp_core::prelude::*;
-use hermes_cosmos_client_components::components::query_client_state::QueryCosmosClientStateFromChainHandle;
+use hermes_cosmos_client_components::impls::query_client_state::QueryCosmosClientStateFromChainHandle;
 use hermes_relayer_components::chain::traits::components::client_state_querier::ClientStateQuerier;
 use hermes_relayer_components::chain::traits::types::client_state::HasClientStateType;
 use ibc_relayer_types::core::ics24_host::identifier::ClientId;
